Add install.sh message

+++ install.sh
@@ -1,3 +1,10 @@
+if [ -z "$1" ]
+  then
+    echo "Please use ./install.sh PLATFORM where platform is one of these:"
+    echo "aix bsd c89 freebsd generic linux macosx mingw posix solaris"
+    exit 0
+fi
+
 curl -O http://www.lua.org/ftp/lua-5.3.1.tar.gz
 tar -xf lua-5.3.1.tar.gz
 rm lua-5.3.1.tar.gz
